+++ p5/src/symbolsTable.c
@@ -201,7 +201,7 @@ bool st_set_scalar_variable(
   String identifier,
   DataType dt,
   Scope scope,
-  int_fast32_t pos
+  int pos
 )
 {
   Symbol *s;
@@ -232,8 +232,8 @@ bool st_set_vector_variable(
   String identifier,
   DataType dt,
   Scope scope,
-  int_fast32_t pos,
-  uint_fast32_t size
+  int pos,
+  uint_fast8_t size
 )
 {
   Symbol *s;
@@ -262,7 +262,7 @@ bool st_set_scalar_parametre(
   SymbolsTable *st,
   String identifier,
   DataType dt,
-  int_fast32_t pos
+  int pos
 )
 {
   Symbol *s;
@@ -286,8 +286,8 @@ bool st_set_vector_parametre(
   SymbolsTable *st,
   String identifier,
   DataType dt,
-  int_fast32_t pos,
-  uint_fast32_t size
+  int pos,
+  uint_fast8_t size
 )
 {
   Symbol *s;
@@ -367,8 +367,8 @@ bool st_insertBlindCurrentScope(
   DataType dataType,
   IdentifierCategory identifierCategory,
   Scope scope,
-  int_fast32_t pos,
-  uint_fast32_t size,
+  int pos,
+  uint_fast8_t size,
   int_fast32_t params,
   int_fast32_t localvars
 )
